% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airwayMetadata}
\alias{airwayMetadata}
\title{Metadata for Airway Experiment}
\format{
A data frame with columns:
\describe{
\item{id}{Sample ID.}
\item{dex}{Whether or not the sample is treated with dex.}
\item{celltype}{Cell line.}
\item{geo_id}{GEO ID.}
}
}
\source{
University of Pennsylvania, United States.
}
\usage{
airwayMetadata
}
\description{
The metadata of an RNAseq experiment characterize the human airway smooth muscle
transcriptome at baseline and treated with a glucocorticosteroid (i.e.
Dexamethasone (Dex), 1μM for 18h).
}
\examples{
\dontrun{
 airwayMetadata
}
}
\references{
Himes et al. (2014). RNA-Seq transcriptome profiling identifies CRISPLD2 as a
glucocorticoid responsive gene that modulates cytokine function in airway smooth
muscle cells. \emph{PloS one, 9}(6), e99625. \href{https://doi.org/10.1371/journal.pone.0099625}{Link}
}
\keyword{datasets}
