% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTogether.R
\name{ClusterTogether}
\alias{ClusterTogether}
\title{Cluster}
\usage{
ClusterTogether(normalized_counts, DE_res, order, tool)
}
\arguments{
\item{normalized_counts}{The count}

\item{DE_res}{The result from DESeq2 or edgeR}

\item{order}{the clustering order}

\item{tool}{tool used for DGE analysis, "DESeq2" or "edgeR"}
}
\description{
This is a function
}
\examples{
# empty

}
\references{
empty
}
