% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMA.R
\name{compMA}
\alias{compMA}
\title{Make MA plots of the genes}
\usage{
compMA(deseq2_result, edger_result, cutoff = 0.05)
}
\arguments{
\item{deseq2_result}{The result from DESeq2, a DESeqResults object.}

\item{edger_result}{A data frame taken from the table component of the returned
value of edgeR topTags function (edgeR::topTags(..)$table).}

\item{cutoff}{padj cutoff. Default: 0.05}
}
\value{
Plots two MA plots side by side, which shows log fold change vs. counts
gotten from DESeq2 and edgeR results respectively.
}
\description{
A function to make MA plots of the genes. The x-axis is the gene counts and
y-axis is log fold change. The gene counts and log fold change are taken from
DESeq2 and edgeR results respectively in the two MA plots shown as output. Plots
are colored by the tools which consider the genes as differentially expressed.
The results are meaningful to compare different tools only when the parameters
used for both tools when performing the DGE analysis are similar.
}
\examples{
# Using airwayCounts and airwayMetadata available with package

# perform DGE analysis with DESeq2
dds <- DESeq2::DESeqDataSetFromMatrix(countData=airwayCounts,
                                      colData=airwayMetadata,
                                      design=~dex,
                                      tidy=TRUE)
dds <- DESeq2::estimateSizeFactors(dds)

dds@colData$dex <- relevel(dds@colData$dex, ref = "control")
dds <- DESeq2::DESeq(dds)

res <- DESeq2::results(dds, tidy=TRUE)
deseq2_res <- DESeq2::lfcShrink(dds, coef = "dex_treated_vs_control",
                                  type = "apeglm")

# perform DGE analysis with edgeR
counts <- data.frame(airwayCounts[,-1], row.names = airwayCounts$ensgene)

diff_list <- edgeR::DGEList(counts, samples = airwayMetadata)

dex <- factor(rep(c("control", "treated"), 4))
design <- model.matrix(~dex)
rownames(design) <- colnames(diff_list)

diff_list <- edgeR::estimateDisp(diff_list, design, robust = TRUE)
fit <- edgeR::glmFit(diff_list, design)
lrt <- edgeR::glmLRT(fit)
edger_res <- edgeR::topTags(lrt, n = dim(lrt)[1])$table

# make the MA plots
compMA(deseq2_result = deseq2_res,
       edger_result = edger_res)

}
