% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEVisComp.R
\name{runDEVisComp}
\alias{runDEVisComp}
\title{Launch Shiny App for DEVisComp}
\usage{
runDEVisComp()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for DEVisComp.
The shiny app permit to plot MA plots, and Volcano plots
comparing the DESeq2 and edgeR input.
}
\examples{
\dontrun{

DEVisComp::runDEVisComp()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
